Organized code, screen resizes, Better Hakkun/camera controls
Code is now more organized and annotated, but some sections are not yet.
This is still a WIP.
Screen resizes are now actively handled in the animate loop
All of Hakkun's body parts and head are now relative to his body rather
than controlled separately.
Hakkun's model now changes angle based on which direction he is facing
It is possible to control the camera independently of Hakkun's position
and angle now.